pyproject.toml
[project]
name = "test-task-in-inno"
version = "0.1.0"
description = ""
authors = [
    {name = "vellimir",email = "kkazancov@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.7"
dependencies = [
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
